June release part 1 (#10)
* FN-DSA (Falcon): Falcon API headers moved to a dedicated public API level; minor coding-guideline and robustness cleanups; further adjustments and justifications based on static code analysis findings.
HQC: Added dedicated HQC API headers; minor coding-guideline related cleanups; additional adjustments based on static code analysis findings.

* Review findings for release resolved

+++ Hqc/Hqc256/FsmSw_Hqc256_parsing.h
@@ -63,9 +63,9 @@
 /* PUBLIC FUNCTION PROTOTYPES                                                                                         */
 /**********************************************************************************************************************/
 
-void FsmSw_Hqc256_Load8_Arr(uint64 *const out64, uint32 outlen, const uint8 *const in8, uint32 inlen);
+void FsmSw_Hqc256_Load8_Arr(uint64 *const out64, uint16 outlen, const uint8 *const in8, uint16 inlen);
 
-void FsmSw_Hqc256_Store8_Arr(uint8 *const out8, uint32 outlen, const uint64 *const in64, uint32 inlen);
+void FsmSw_Hqc256_Store8_Arr(uint8 *const out8, uint16 outlen, const uint64 *const in64, uint16 inlen);
 
 void FsmSw_Hqc256_Secret_Key_To_String(uint8 *const sk, const uint8 *const sk_seed, const uint8 *const sigma,
                                        const uint8 *const pk);
